refactor(sut): move metrics to common directory
