+++ src/sut/dumblog/bench/Common.hs
@@ -40,10 +40,10 @@ rEAD_FREQUENCY :: Int
 rEAD_FREQUENCY = 80
 
 nUM_OF_CLIENTS :: Int
-nUM_OF_CLIENTS = 5000
+nUM_OF_CLIENTS = 20
 
 iTERATIONS :: Int
-iTERATIONS = 50
+iTERATIONS = 1000
 
 vALUE_TO_WRITE :: ByteString
 vALUE_TO_WRITE = LBS.pack "Dumblog"
